more updates to template generator
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5342 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/ui/editors/EntityEditor.java
@@ -838,8 +838,8 @@ public class EntityEditor extends EditorPart implements MapPart{
 		this.isEditMode = isEdit;
 		if (entity != null) initControl(entity);
 	
-		deleteItem.setEnabled(isEdit);
-		editItem.setSelection(isEdit);
+		if (!deleteItem.isDisposed()) deleteItem.setEnabled(isEdit);
+		if (!editItem.isDisposed()) editItem.setSelection(isEdit);
 	}
 	
 	public void setDirty(boolean isDirty){
